+++ web/src/routes/(app)/courses/+page.svelte
@@ -34,7 +34,7 @@
 </header>
 
 {#if form?.message}
-	<p class="banner banner-bad mb-4" role="alert">{form.message}</p>
+	<p class="banner-bad mb-4" role="alert">{form.message}</p>
 {/if}
 
 {#if adding}